Send PMSR indications from notify.c to AIDL
Bug: 476895058
Test: m
CRs-Fixed: 4622308
(cherry picked from commit fc971d7eebff3a9294ae78d89077313e1ff3cb44)

+++ wpa_supplicant/aidl/supplicant_wifi_rtt_controller.cpp
@@ -530,8 +530,8 @@ ndk::ScopedAStatus SupplicantWifiRttController::rangeCancelInternal(
 	AidlManager *aidl_manager = AidlManager::getInstance();
 	if (aidl_manager) {
 		aidl_manager->notifyRttContinuousRangingTerminatedEvent(wpa_s, 
-			ISupplicantWifiRttControllerEventCallback::
-			ContinuousRangingTerminateReasonCode::USER_REQUEST);
+			static_cast<u32>(ISupplicantWifiRttControllerEventCallback::
+			ContinuousRangingTerminateReasonCode::USER_REQUEST));
 	}
 	current_cmd_id_ = -1;
 	return ndk::ScopedAStatus::ok();
